+++ src/extension/app/components/action-bar/action-bar.js
@@ -19,10 +19,10 @@ export class ActionBar extends LitElement {
     .action-bar {
       display: flex;
       border-radius: var(--spectrum2-sidekick-border-radius);
-      color: var(--spectrum2-sidekick-color));
+      clip-path: border-box;
+      color: var(--spectrum2-sidekick-color);
       background-color: var(--spectrum2-sidekick-background);
       border: 1px solid var(--spectrum2-sidekick-border-color);
-      box-shadow: var(--sidekick-box-shadow);
       backdrop-filter: var(--sidekick-backdrop-filter);
       -webkit-backdrop-filter: var(--sidekick-backdrop-filter);
     }
@@ -59,6 +59,7 @@ export class ActionBar extends LitElement {
       .action-bar {
         border-radius: 0;
       }
+    }
   `;
 
   render() {