+++ src/extension/app/store/app.js
@@ -355,7 +355,9 @@ export class AppStore {
           let processedUrl;
           if (url) {
             const target = new URL(url, `https://${innerHost}/`);
-            target.searchParams.set('theme', this.theme);
+            if (isPalette || isPopover) {
+              target.searchParams.set('theme', this.theme);
+            }
             if (passConfig) {
               target.searchParams.append('ref', this.siteStore.ref);
               target.searchParams.append('repo', this.siteStore.repo);
@@ -637,7 +639,7 @@ export class AppStore {
   isSharePointEditor(url) {
     const { pathname, searchParams } = url;
     return this.isSharePoint(url)
-      && pathname.match(/\/_layouts\/15\/[\w]+.aspx/)
+      && pathname.match(/\/_layouts\/15\/[\w]+\.aspx/)
       && (searchParams.has('sourcedoc') || searchParams.has('id'));
   }
 
@@ -656,6 +658,21 @@ export class AppStore {
   }
 
   /**
+   * Checks if a content source URL is Document Authoring (DA).
+   * DA content source URLs start with a <code>markup:</code> prefix.
+   * @param {string} contentSourceUrl The content source URL string to check
+   * @returns {boolean} <code>true</code> if URL is DA, else <code>false</code>
+   */
+  isDA(contentSourceUrl) {
+    try {
+      const { hostname } = new URL(contentSourceUrl.substring(7));
+      return hostname.endsWith('.da.live');
+    } catch (e) {
+      return false;
+    }
+  }
+
+  /**
    * Returns a label for the content source
    * @returns {string} The content source label
    */
@@ -668,7 +685,11 @@ export class AppStore {
     } else if (sourceLocation?.startsWith('gdrive:')) {
       return 'Google Drive';
     } else if (sourceLocation?.startsWith('markup:')) {
-      return contentSourceEditLabel || 'BYOM';
+      if (contentSourceEditLabel) {
+        return contentSourceEditLabel;
+      } else {
+        return this.isDA(sourceLocation) ? 'Document Authoring' : 'BYOM';
+      }
     } else if (contentSourceType === 'onedrive') {
       return 'SharePoint';
     } else if (contentSourceType === 'google') {