+++ CHANGELOG.md
@@ -1,3 +1,127 @@
+## [7.25.4](https://github.com/adobe/aem-sidekick/compare/v7.25.3...v7.25.4) (2025-12-17)
+
+
+### Bug Fixes
+
+* move project admin to tools ([#736](https://github.com/adobe/aem-sidekick/issues/736)) ([ab11423](https://github.com/adobe/aem-sidekick/commit/ab11423d1c06c476b63ce92678c518358228b837))
+
+## [7.25.3](https://github.com/adobe/aem-sidekick/compare/v7.25.2...v7.25.3) (2025-12-09)
+
+
+### Bug Fixes
+
+* localize tooltip and activate copy all button for localized column names for path/url/href/link ([ad9d866](https://github.com/adobe/aem-sidekick/commit/ad9d86602296c593f42df5c871807b2101cd6fc6))
+
+## [7.25.2](https://github.com/adobe/aem-sidekick/compare/v7.25.1...v7.25.2) (2025-12-08)
+
+
+### Bug Fixes
+
+* ensure relative image urls are handled correctly ([#734](https://github.com/adobe/aem-sidekick/issues/734)) ([f823b5e](https://github.com/adobe/aem-sidekick/commit/f823b5edefe5b732f94d42a14c2380c07ccfdbd7))
+
+## [7.25.1](https://github.com/adobe/aem-sidekick/compare/v7.25.0...v7.25.1) (2025-12-05)
+
+
+### Bug Fixes
+
+* closed sidekick can crash page when being resized ([#739](https://github.com/adobe/aem-sidekick/issues/739)) ([6348766](https://github.com/adobe/aem-sidekick/commit/6348766e946d9ee67a9a583add3cd8045a0ca897))
+* **regression:** add back border radius ([c8754e9](https://github.com/adobe/aem-sidekick/commit/c8754e96fc319acdd08cc4356cf9e4b0a60b7c4c))
+* restore box-shadow visibility for action bar ([#733](https://github.com/adobe/aem-sidekick/issues/733)) ([eaf05aa](https://github.com/adobe/aem-sidekick/commit/eaf05aa7860d11c5e665619a6dccf180e57c3168))
+
+# [7.25.0](https://github.com/adobe/aem-sidekick/compare/v7.24.1...v7.25.0) (2025-11-18)
+
+
+### Features
+
+* **env-switcher:** show source item for DA by default ([#731](https://github.com/adobe/aem-sidekick/issues/731)) ([d868c8f](https://github.com/adobe/aem-sidekick/commit/d868c8f33d4e776f0d57ebe1ef60f2d813c8f965))
+
+## [7.24.1](https://github.com/adobe/aem-sidekick/compare/v7.24.0...v7.24.1) (2025-11-17)
+
+
+### Bug Fixes
+
+* 429 from backend on status request shows generic error message ([#728](https://github.com/adobe/aem-sidekick/issues/728)) ([4396e27](https://github.com/adobe/aem-sidekick/commit/4396e27f5b2d72d8a5c75f7a32eed804a2801f33))
+
+# [7.24.0](https://github.com/adobe/aem-sidekick/compare/v7.23.3...v7.24.0) (2025-11-13)
+
+
+### Features
+
+* ability to reposition sidekick ([#720](https://github.com/adobe/aem-sidekick/issues/720)) ([7a44e38](https://github.com/adobe/aem-sidekick/commit/7a44e383484d422c8ba1985be7b9979c3848042c)), closes [#714](https://github.com/adobe/aem-sidekick/issues/714)
+
+## [7.23.3](https://github.com/adobe/aem-sidekick/compare/v7.23.2...v7.23.3) (2025-11-12)
+
+
+### Bug Fixes
+
+* popover iframe should allow clipboard write ([#726](https://github.com/adobe/aem-sidekick/issues/726)) ([6406545](https://github.com/adobe/aem-sidekick/commit/6406545ffbc934ffc85ae736cfe1844063ec625b))
+
+## [7.23.2](https://github.com/adobe/aem-sidekick/compare/v7.23.1...v7.23.2) (2025-11-10)
+
+
+### Bug Fixes
+
+* revert sharepoint lookup optimization ([#722](https://github.com/adobe/aem-sidekick/issues/722)) ([e7aafd4](https://github.com/adobe/aem-sidekick/commit/e7aafd44da9cdca5e9e6ed230505c34c0a98d3df))
+
+## [7.23.1](https://github.com/adobe/aem-sidekick/compare/v7.23.0...v7.23.1) (2025-11-10)
+
+
+### Bug Fixes
+
+* improve error messages ([#719](https://github.com/adobe/aem-sidekick/issues/719)) ([d1b0c12](https://github.com/adobe/aem-sidekick/commit/d1b0c12010c85aa0940669991e07a63cfbada1c4))
+* improve error messages ([#719](https://github.com/adobe/aem-sidekick/issues/719)) ([6cbe8bc](https://github.com/adobe/aem-sidekick/commit/6cbe8bcea0e14b11f0f5ff91857cb42eb1ec432f))
+
+# [7.23.0](https://github.com/adobe/aem-sidekick/compare/v7.22.2...v7.23.0) (2025-11-06)
+
+
+### Bug Fixes
+
+* copy icon with sorting alignment and minor fixes ([a6be3fa](https://github.com/adobe/aem-sidekick/commit/a6be3fa721aaae7664edaa9e7c4499df14c268e5))
+* sorting when no columns obj available ([233ffa7](https://github.com/adobe/aem-sidekick/commit/233ffa70260348caedabcb77a258ca03a59d04ab))
+
+
+### Features
+
+* add copy button to url/path column ([5ba6c01](https://github.com/adobe/aem-sidekick/commit/5ba6c0168d3d3ece9ef7560ba19ee6c129160bd4))
+* add green check confirmation on copy ([81948d4](https://github.com/adobe/aem-sidekick/commit/81948d4bc4fa63ead774d183eea84c99ebcf1f47))
+
+## [7.22.2](https://github.com/adobe/aem-sidekick/compare/v7.22.1...v7.22.2) (2025-10-23)
+
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#703](https://github.com/adobe/aem-sidekick/issues/703)) ([ab5e96d](https://github.com/adobe/aem-sidekick/commit/ab5e96ddf9a666a9adc0590b1b56558d15af5a4d))
+
+## [7.22.1](https://github.com/adobe/aem-sidekick/compare/v7.22.0...v7.22.1) (2025-10-23)
+
+
+### Bug Fixes
+
+* **login:** display email only if no name ([#707](https://github.com/adobe/aem-sidekick/issues/707)) ([d9bdc73](https://github.com/adobe/aem-sidekick/commit/d9bdc7383439ada3c0b78f033f3072e1de72aa65)), closes [#708](https://github.com/adobe/aem-sidekick/issues/708)
+
+# [7.22.0](https://github.com/adobe/aem-sidekick/compare/v7.21.0...v7.22.0) (2025-10-22)
+
+
+### Features
+
+* add resize popover and palette APIs ([#705](https://github.com/adobe/aem-sidekick/issues/705)) ([573c464](https://github.com/adobe/aem-sidekick/commit/573c46490ffef34953e58764008d8caec7c0cb7d))
+
+# [7.21.0](https://github.com/adobe/aem-sidekick/compare/v7.20.1...v7.21.0) (2025-10-21)
+
+
+### Features
+
+* add palette and popover close APIs ([#704](https://github.com/adobe/aem-sidekick/issues/704)) ([9ab6ed6](https://github.com/adobe/aem-sidekick/commit/9ab6ed64d17afd98a582d0c757c43ddeca9c9b48))
+
+## [7.20.1](https://github.com/adobe/aem-sidekick/compare/v7.20.0...v7.20.1) (2025-10-08)
+
+
+### Bug Fixes
+
+* **auth:** allow localhost during auth CORS header config ([4834f85](https://github.com/adobe/aem-sidekick/commit/4834f85569f4b3c1a9049c211aed38bb6af1bad3))
+* **auth:** simplify localhost regex to port 3000 only ([4cedfee](https://github.com/adobe/aem-sidekick/commit/4cedfee84ab1a10e9a93253d0543a4ff9ea846e3))
+* **auth:** update test to match localhost regex pattern ([36fd793](https://github.com/adobe/aem-sidekick/commit/36fd7934828a8250db5f351656609724b1d0cd15))
+
 # [7.20.0](https://github.com/adobe/aem-sidekick/compare/v7.19.7...v7.20.0) (2025-10-03)